:hammer: Rename converToExtendedNquads to convertToExtendedJsonLd
