+++ src/api/exporters/exportExtendedNquads.js
@@ -17,8 +17,8 @@ const exporter = (config, fields, characteristics, stream) =>
                 ),
             }),
         )
-        .pipe(ezs('convertToExtendedNquads', { config }))
-        .pipe(ezs('debug'))
+        .pipe(ezs('convertToExtendedJsonLd', { config }))
+        // .pipe(ezs('debug'))
         .pipe(ezs('convertJsonLdToNQuads'));
 
 exporter.extension = 'nq';
