+++ src/api/statements/index.js
@@ -9,7 +9,7 @@ import JSONLDObject from './JSONLDObject';
 import extractIstexQuery from './extractIstexQuery.js';
 import scroll from './scroll.js';
 import convertJsonLdToNQuads from './convertJsonLdToNQuads.js';
-import convertToExtendedNquads from './convertToExtendedNquads.js';
+import convertToExtendedJsonLd from './convertToExtendedJsonLd.js';
 import convertToAtom from './convertToAtom.js'
 import LodexContext from './LodexContext.js'
 import LodexConfig from './LodexConfig.js'
@@ -30,7 +30,7 @@ export default {
     extractIstexQuery,
     scroll,
     convertJsonLdToNQuads,
-    convertToExtendedNquads,
+    convertToExtendedJsonLd,
     convertToAtom,
     LodexContext,
     LodexConfig,
